Alot of different changes
Check the diff lol. Added some JS to change the color of the background for each thumbnail, as well as overlay the skills for each project over the thumbnail. Introduced placeholders for upcoming projects. Modal box now pops up when clicking the read more button. Changed the Contact Me section so that it's centered. All around wellness changes for the responsiveness of the site. I think that's it.