Changes to cease all other animations until current one finishes
Set a variable 'youShallNotRun' that keeps track of whether an animation is running for the skills section.  If it's true, then nothing else runs. Once animation stops then value goes back to false and other animations can run.

Still need to make sure that the animation for css has the proper timing.