Changes in the Skills section
On hover, all of the Skills circles say 'Click Me'. Added popover for the CSS3 skill after the first click so that it says it won't work again.  Made the animations activate on click instead of hover. Started setting up the animation between JS and WordPress circles -- namely, setup the popovers for the WordPress circle.