Changes to background icons and Skills sections circles.
The background icons now spin at a slower rate.  Also changed the positioning of the icons to spread them out more evenly and added an additional one.  On hover, the Skills section circles now get filled in with a color that matches the border and the text color turns to white.